Revert "refactor: Make use of `value class` for `SearchHistoryId`"
This reverts commit 5351be0a6d8fbe7df6dfe03ce3624283ee38e94f.

+++ app/src/main/kotlin/com/prajwalch/torrentsearch/data/local/dao/SearchHistoryDao.kt
@@ -18,7 +18,7 @@ interface SearchHistoryDao {
     fun observeAll(): Flow<List<SearchHistoryEntity>>
 
     @Query("DElETE from search_history where id=:id")
-    suspend fun deleteById(id: Int)
+    suspend fun deleteById(id: Long)
 
     @Query("DELETE from search_history")
     suspend fun deleteAll()
